pyproject.toml
[project]
name = "dbt-tpch"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "dbt-duckdb>=1.10.0",
]
